Make vim-airline respect an existing tabline
This Fixes #833.
Set :let airline#extensions#tabline#keep_orig_tabline = 1 in your vimrc
to make vim-airline try to evaluate the existing tabline setting and
only beautify it (e.g. do not rename the labels, etc.). This works by
evaluating and parsing the current tabline setting, so this is actually
a little bit fragile and might break for some plugins. Therefore this is
only an experimental feature.